requirements.txt
Flask==3.0.0
google-api-python-client==2.111.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.0
